Error messages should generally include throwing function's name
There may be a couple corner cases which are not currently handled
correctly, because they aren't using the GoError function to
check/convert errors.  When logging an error message GoError should be
used to convert the LVal to an ErrorVal which also provides the
WriteTrace() methed for dumping stack traces.