Facilitate more advanced error handling in handler-bind
The 'condition symbol can catch multiple types of errors (it is the only
type of error inheritence that currently exists).  So handler-bind needs
to tell the handler what error condition symbol matched the evaluated
handler.

For example, consider an advanced macro

    (defmacro assert-condition (condition-symbol expression)

The macro generates error handling code which needs to check for the
given condition, print the condition, and print other encountered
conditions if they do not match the given condition-symbol.

In order to allow comparison of symbols without *always* converting them
to strings builtin functions ``symbol?'' and ``symbol='' have been
added.