Trigger a runtime error when encountering (quasiquote (unquote-splicing ...))
The result of such a quasiquote expression is not well defined.  In
general unquote-splicing is not valid outside of a list expression but
the quasiquote does not count as a list expression because it may not
actually produce a list as output.

    (quasiquote 1)  ; evaluates to '1 (which is equivalent to 1)

The error message is taken from racket, and it implies that there may be
other contexts in which unquote-splicing is invalid.  But I have not yet
thought of any other examples of invalid contexts.

Here are some examples from racket:

    Welcome to Racket v6.12.
    > (quasiquote (if 123 (unquote-splicing '(1 2))))
    '(if 123 1 2)
    > (quasiquote (if 123 '(unquote-splicing '(1 2 3))))
    '(if 123 (quote 1 2 3))

All this should be OK in elps as well.