Builtin function ``rest'' accepts arrays, unlike cdr
Rest always returns a list even for array inputs.  It seems like it
might be best to copy data out of as array, though for now it doesn't do
this.  The cdr function doesn't do this.

This is something that needs to worked out in the embedding guide.  It's
not consistently handled, and generally unclear what function owns any
given LVal, and which LVals may have their data modified.