A basic regexp package -- change ``format-string'' handling of LString values
Pereviously, format-string would not differentiate between arguments in
the following expressions:

    (format-string "hello {}" "jake")
    (format-string "hello {}" (quote "jake"))

But it was not possible for format-string to render a string in its
"source representation" (surrounded by double quotes, with special
characters escaped).

Now, the two expressions will render different strings:

    hello, jake
    hello, "jake"

This seemed to be quite useful for assertion macros which render custom
failure messages (that are rendered using the format-string builtin).
It didn't seem as sensible to keep the ``format-string'' semantics and
define a new ``quote-string'' function which turn a string into a string
containing the argument's quoted representation.