The `append` and `append!` functions work with bytes
Also add new functions `append-bytes` and `append-bytes!` which are
primarily defined for the purpose of efficiently appending strings and
other `bytes` values onto a `bytes` value -- where the append and
append! functions append invidual byte values (integers).

The internal representation for `bytes` has changed to use the
LVal.Native field, and the byte slice has been made made into an
indirect reference.  The use of Native should very slightly speed up
LVal.Copy by reducing the size of an LVal by the size of 3 words
(24-bytes), as a slice is represented as 3 machine words <pointer>,
<len>, and <cap>.

Changing the stored value to `*[]byte` instead of `[]byte` allows
append! and append-bytes! to work.  The reasoning here is the same as
what prompted the change in internal representation for vectors a short
time ago.