Change the debug representation of `bytes` values to avoid confusing
The old debug representation would not produce a desirable result when
copy-pasted back into the repl.  The new representation will not return
a `bytes` type but it won't return anything else either.

The sequence #< is defined to always produce a parse error when
encountered in common-lisp.  So values which do not have a "literal"
representation in lisp should generally be stringified with the prefix
`#<`.